API

Parquet.BatchedColumnsCursor

Parquet.ColCursor

Parquet.Dataset

Parquet.DatasetPartitions

Parquet.File

Parquet.Page

Parquet.PageLRU

Parquet.Schema

Parquet.Table

Parquet.TablePartition

Parquet.TablePartitions

Parquet.read_parquet

Parquet.write_parquet

Basic Usage

Parquet.File — Type

Parquet.File(path; map_logical_types) => Parquet.File

Represents a Parquet file at path open for reading. Options to map logical types can be provided via map_logical_types.

map_logical_types can be one of:

false: no mapping is done (default)

true: default mappings are attempted on all columns (bytearray => String, int96 => DateTime)

A user supplied dict mapping column names to a tuple of type and a converter function

Returns a Parquet.File type that keeps a handle to the open file and the file metadata and also holds a LRU cache of raw bytes of the pages read.

Parquet.Table — Type

Parquet.Table(path; kwargs...)

Returns the table contained in the parquet file in a Tables.jl compatible format.

Options:

rows: The row range to iterate through, all rows by default.

batchsize: Maximum number of rows to read in each batch (default: row count of first row group).

use_threads: Whether to use threads while reading the file; applicable only for Julia v1.3 and later and switched on by default if julia processes is started with multiple threads.

One can easily convert the returned object to any Tables.jl compatible table e.g. DataFrames.DataFrame via

using DataFrames
df = DataFrame(read_parquet(path))

Parquet.Dataset — Type

Parquet.Dataset(path; kwargs...)

Returns the table contained in the parquet dataset in an Tables.jl compatible format. A dataset comprises of multiple parquet files and optionally some metadata files.

These options if provided are passed along while reading each parquet file in the dataset:

filter: Filter function that takes the path to partitioned file and returns boolean to indicate whether to include the partition while loading. All partitions are loaded by default.

batchsize: Maximum number of rows to read in each batch (default: row count of first row group). Applied to each file in the partition.

use_threads: Whether to use threads while reading the file; applicable only for Julia v1.3 and later and switched on by default if julia processes is started with multiple threads.

column_generator: Function to generate a partitioned column when not found in the partitioned table. Parameters provided to the function: table, column index, length of column to generate. Default implementation determines column values from the table path.

One can easily convert the returned object to any Tables.jl compatible table e.g. DataFrames.DataFrame via

using DataFrames
df = DataFrame(read_parquet(path))

Parquet.read_parquet — Function

read_parquet(path; kwargs...)

Returns the table contained in the parquet file or dataset (partitioned parquet files in a folder) in a Tables.jl compatible format.

Options:

rows: The row range to iterate through, all rows by default. Applicable only when reading a single file.

filter: Filter function to apply while loading only a subset of partitions from a dataset.

batchsize: Maximum number of rows to read in each batch (default: row count of first row group). Applied only when reading a single file, and to each file when reading a dataset.

use_threads: Whether to use threads while reading the file; applicable only for Julia v1.3 and later and switched on by default if julia processes is started with multiple threads.

column_generator: Function to generate a partitioned column when not found in the partitioned table. Parameters provided to the function: table, column index, length of column to generate. Default implementation determines column values from the table path.

One can easily convert the returned object to any Tables.jl compatible table e.g. DataFrames.DataFrame via

using DataFrames
df = DataFrame(read_parquet(path))

Parquet.write_parquet — Function

write_parquet(io, tbl; compression_codec="SNAPPY")

Write a parquet file from a Tables.jl compatible table e.g DataFrame Write table (Table.jl compatible object) as a parquet to io. io can either be an IO stream object or a string, in which case it will be written to the file name given.

Available compression codecs are: "SNAPPY", "UNCOMPRESSED", "ZSTD", "GZIP".

Low-level Usage

Parquet.Page — Type

Page

Data structure representing a parquet "page".

Column chunks are divided into pages which are conceptually individual units in terms of compression and encoding. Multiple page types can be contained in a simble column chunk.

Parquet.PageLRU — Type

PageLRU

Keeps a cache of pages read from a file. Pages are kept as weak refs, so that they can be collected when there's memory pressure.

Parquet.TablePartition — Type

TablePartition

Represents one partition of the parquet file. Typically a row group, but could be any other unit as mentioned while opening the table.

Parquet.TablePartitions — Type

TablePartitions

Iterator to iterate over partitions of a parquet file, returned by the Tables.partitions(table) method. Each partition is typically a row group, but could be any other unit as mentioned while opening the table.

Parquet.ColCursor — Type

ColCursor{T}

Column cursor iterates through all values of the column, including null values. Each iteration returns the value (as a Union{T,Nothing}), definition level, and repetition level for each value. Row can be deduced from repetition level.

Parquet.BatchedColumnsCursor — Type

BatchedColumnsCursor

Create cursor to iterate over batches of column values. Each iteration returns a named tuple of column names with batch of column values. Files with nested schemas can not be read with this cursor.

Constructors

BatchedColumnsCursor(par::Parquet.File; kwargs...)

Arguments

rows: the row range to iterate through, all rows by default.

batchsize: maximum number of rows to read in each batch (default: row count of first row group).

reusebuffer: boolean to indicate whether to reuse the buffers with every iteration; if each iteration processes the batch and does not need to refer to the same data buffer again, then setting this to true reduces GC pressure and can help significantly while processing large files.

use_threads: whether to use threads while reading the file; applicable only for Julia v1.3 and later and switched on by default if julia processes is started with multiple threads.

Parquet.DatasetPartitions — Type

DatasetPartitions

Iterator to iterate over partitions of a parquet dataset, returned by the Tables.partitions(dataset) method. Each partition is a Parquet.Table.

Parquet.Schema — Type

Schema

Parquet table schema.